Update pagerduty type to use api_key
when I get a pagerduty integration via the api with `GET api/integration/pager_duty/<id>`, it doesn't show service_key, but api_key. Setting it manually in the ui makes api_key show up, and repeatedly applying terraform with a service_key does not update the integration in the console.